Grids: extract Cache tests into a separate file
